[bt][settings] Fix crash when growing discovery list
The code used to crash during device discovery while trying to grow a
fixed-size list object. This was because the initially constructed
growable list was getting reassigned to a FIDL API response upon
refresh, which returns a fixed-size list.

Fixed this by always operating on the same growable list.

Also added "appearance" as a debug-mode field.

Bug: none
Test: 1. Build workstation.x64 product and run on a NUC.
      2. Log-in as Guest.
      3. Type "settings" in the search box.
      4. Navigate to the "Bluetooth" tab.
      5. This will start populating discovery results immediately. The
      app shouldn't crash.
Change-Id: Ife2830a673040018e31f3d6a1576aebd627b4ace