[benchmarks] Add flutter-driver option to scroll-flutter
Previously, flutter_driver support was added using
flutter_driver_extendable in BUILD.gn. However, this is only effective
on debug builds of Fuchsia. We want to be able to use this for automated
tests in release builds, so replace it with an explicit command-line
argument that enables flutter_driver in the scroll-flutter app and set
product = false in BUILD.gn to force a non-release Flutter build.

Bug: 55832

Test: Manually tested in FEMU with a WIP benchmark; this appears to
allow it to use flutter_driver even on release builds of Fuchsia, where
previously it was only usable on debug builds. Removing the
--flutter-driver argument when launching the app causes the attempt to
use flutter_driver to fail in the same way the previous version did on
release builds.

Change-Id: If75287205774996eb27921147e72c6984a8bca69
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/409634
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Testability-Review: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Sean Gillen <sdgillen@google.com>