[settings] Migrate back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Bug: 55719
Test: fx run-host-tests ermine_settings_unittests

Change-Id: I5e48f8eca20200c351325b0cb693c8a74c097d79
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/414055
Reviewed-by: William Leshner <wleshner@google.com>
Testability-Review: Chip Fukuhara <cfukuhara@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>