[ermine] Remove deprecated use of ViewRefInstalled service.
Scenic use to require using this service to check if a view was
connected to the view tree, before requesting focus on it. This
restriction was removed recently. This change removes the usage of this
service from Ermine.

Fixed: 61946
Change-Id: I822b35a86358b11693469ebf293050d9989283d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/523500
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Yeonhee Lee <yhlee@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>