[Workstation][Performance] Add tab opening navigation performance test
This CL adds the following navigation tests:
- add tabs

More navigation tests (rearranging tabs, closing tabs) to follow. This CL does not visualize data to a dashboard; this will be addressed in subsequent CLs. This CL simply traces the navigation test scenarios and formats the data to be utilized in a dashboard.

The performance tests will be added to workstation.gni later in fxr/518621.

Due to fxb/76046, simple browser crashes on its second launch. This is low risk because the performance tests are not included in the workstation product yet and tests can be run individually (locally) until this bug is addressed.

Test:
(cherry-pick fxr/518621)
fx test -o --e2e experiences_ermine_session_shell_performance_test

Bug: 68516

Change-Id: If11ed753d04cc3fe91dd6035d207a74b71849c56
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/526582
Commit-Queue: Charles Whitten <cwhitten@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>