[ermine] Allow dismissing screen saver on keyboard activity
This change fixes a corner case of dismissing the screen saver when a
child view is running.

The screen saver can be dismissed using pointer or keyboard activity.
But this does not work when a child view has focus, the shell does not
receive keyboard events then.

In order to receive keyboard events, while the screen saver is active,
this change moves the focus to the shell view by showing the overlays.
This ensures that we get keyboard events while child view is running.

Change-Id: I75eaad47a1228087f4926691bb8b1cd6b0aeea17
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/552640
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>