[ermine] Fix view hittest after app switcher selects it
This change fixes the issue where the view selected using the app
switcher, does not respond to pointer clicks. This is because the app
switcher is an overlay and we disable hittest when the switcher is
visible. But it is not enabled correctly when the switcher is dismissed.
It needs to be dismissed when the target view has received focus.

Also fixes an exception thrown in channel_service, where we were reusing
an already bound proxy reference.

Fixed: 81594,81358
Change-Id: I9625d8f7b1ebbcf685fe2b7259459690723dae81
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/561981
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>