[e2e] Deflake 'screen is not black' test on ermine
The test is flaking because of a race condition when the screen is
rendered and the screenshot taken. The intention is to ensure that the
screen is rendered before the screenshot is taken. This change uses
flutter driver's screeshot method before taking scenic's screenshot to
increase the odds.

Bug: 79921

Change-Id: I72ddbe064355e36aa46fabbef2b9ad4cb87c568d
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/567561
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>