[ermine] Remove mobx from AppState
MobX observables can be moved out of AppState, which is an abstract
class. It is then an implementation detail, handled in
app_state_impl.dart and hidden behind getters/setters.

This also has the benefit of removing all '.value' dereferences, which
were leaking the implementation detail that object is an observable.
Instead of exposing MobX Actions, we just provide simple methods that
are more typesafe because, unlike Actions, their parameters can be
passed directly instead of in an Array.

ViewState and SettingsState will be updated in subsequent cls.

Test: All updated unittests pass.
Change-Id: I83ebceeb9be2e60d8de4d37276154b79c867c02a
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/574153
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Reviewed-by: Charles Whitten <cwhitten@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>