[Ermine][Quicksettings] Add initial WiFi UI
Adds initial WiFi quicksetting UI to view scanned networks. As a temporary solution quicksetting is disabled until the scanned networks list is populated. This has two purposes: the first one is to prevent the quicksetting appearing active before scanning is available on device, as well as prevents the user from accessing the wifi quicksetting until after the initial scan has taken place (takes around 20 seconds initially). Networks that are not compatible are disabled/greyed out. Note that there is a bug where network names with non-ASCII characters may appear in the scanned networks list and appear as blank networks in the UI. This will be addressed in a subsequent CL.

Bug: 79885

Change-Id: I68ae7ebd51ff2e3d26f990bf217fd5307f7e58a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/592913
Commit-Queue: Charles Whitten <cwhitten@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>