[ermine] Launch OOBE from workstation session
The workstation session launches OOBE at start. OOBE then launches the
ermine user shell immediately if OOBE was completed. Otherwise it shows the
OOBE UX and launches ermine at the end.

At present, OOBE experence defaults to OFF and launches Ermine shell.
This keeps the current user flow. It will be turned ON in the future
when we have user consent and login flow implemented.

This is in preparation of user authentication and login shell.

Bug: 81600
Change-Id: I6d7ecfc4282ffbb18a7dc9fc0e5e4687432e2bb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/564514
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Reviewed-by: Charles Whitten <cwhitten@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>