[ermine] Remove peek from right edge to allow scrollbar access
Ermine allows displaying the sysui using pointer peek on the left and
right edge of the screen when apps are fullscreen. But this interferes
with the user trying to use the scrollbar on apps like Chromium that are
also on the right edge of the screen.

We still keep the peek behavior on left edge, as well as the keyboard
shortcut (Meta key) to display sysui when apps are fullscreen.

Change-Id: Ide606531e8495a982c0ca5b7fce66f34dc88d29b
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/597147
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Reviewed-by: Charles Whitten <cwhitten@google.com>
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>