[e2e][simple-browser] Converted ermine_testserver to a flutter component
ermine_testserver was launched using Component.launch in e2e test and therefore, it did not have any views. However, closing the test server through the workaround using a special URL turned out to be very flaky in testing environment. For this reason, this CL converts the ermine_testserver component to a flutter component with a view so that it can be closed by closing its view. Verified that the test was much more stable and passed ten times in a row on local machine.

Bug: 80345

Change-Id: I609f5a0271d941a16f4c1616a663dcaf0ba3bd92
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/596406
Commit-Queue: Yeonhee Lee <yhlee@google.com>
Reviewed-by: Charles Whitten <cwhitten@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>