[workstation_session] Define a storage capability.
This defines a storage capability that workstation session may offer to
components in the session. The storage capability is backed by a directory
in the outgoing path that is populated when workstation_session
successfully retreives a directory from the account system.

Testing: The backend work for this is not complete yet, but I did test
  a variant of this CL that late bound the outgoing directory using
  an isolated persistent storage input. The storage example component
  was able to successfully read and write using the storage capability.

Bug: 81601
Change-Id: Ieb1f27c6c543e6244daa2f3a947f7155bf9e0de4
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/598654
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>