[Ermine][Quicksettings] Add forget to WiFi UI
This CL introduces the ability to forget saved networks. To forget a network, press the ‘more’ icon on the far right of a saved network entry to view the ‘forget’ popup menu. Click ‘forget’ to remove the corresponding network. There is no UI indication of a failed forget attempt; this will be added in a subsequent CL. Tested on Atlas.

Bug: 79885

Change-Id: I7fe688f1a4efadf94084120d24b47aab6642aac7
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/602231
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Commit-Queue: Charles Whitten <cwhitten@google.com>