[login] Wire login UX with AccountManager
This change wires up the password authentication logic with the login
UX. It moves the authentication logic from workstation_session to login
shell. The login shell implements the create account with password
functionality and subsequently login in using that password. The ermine
user shell implements logout using a keyboard shortcut.

The ability to mount the encrypted partition as "account_data" directory
is still blocked on http://fxb/89690 and fixes to flutter runner.

Although everything is wired up and functional (except for the flutter
runner fix), it is behind a build flag 'start_oobe', which is set to
false. In addition, the code that enables this is also hard disabled
until we are ready to land the user data protection feature.

Fixed: 85564,85587,87678,87679
Bug: 85576

Change-Id: I7c19aea144bf0df73aaa2cb46f053fc2d8949589
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/614005
Reviewed-by: Charles Whitten <cwhitten@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>