[ermine] Dismiss screensaver on Esc key shortcut
After change to the flatland compositor, it is not possible to dismiss
screensaver on ermine using keyboard or mouse input. This is because,
ermine does not get those event when a child view (screensaver) is
running fullscreen.

This change allows dismissing the screensaver using the Esc keyboard
shortcut, which is received by ermine.

Bug: 93800
Change-Id: I273c1c61984117a4777581d303dd4b1ef2297d6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/645963
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>