[input] Prepare fuchsia.ui.focus.FocusChainListenerRegistry for rerouting
We need to remove one of two uses of `f.ui.f.FCLR` from Ermine.  The
cutover needs to be atomic, and needs to cut only one of the two uses
in Ermine.

For this reason, I'm first decoupling the two routes.  Followup changes
will untangle the two uses by having one consume the newly-introduced
`*-workstation` service name.

Once the uses are untangled, we can make the cutover atomically in
`fuchsia.git`.  This ensures that in no commits the system is left
without the focus distribution code that used to be provided by Ermine.

Once the above is done, we can dismantle the unneded code in Ermine, and
remove the duplicate routing.

See: https://fuchsia-review.googlesource.com/c/fuchsia/+/666850

Bug: 95901
Change-Id: I498aef259ceb382d41efed623b5db63844d6ce9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/666329
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>