[input] Connects main.dart to an alternate service
We are migrating focus distribution away from Ermine.  Focus
distribution is the responsibility of the input pipeline, which ended up
in Ermine by necessity, the reasons for which no longer exist.

We're now pulling focus distribution to where it's supposed to be,
see:
https://fuchsia-review.googlesource.com/c/fuchsia/+/666844

To do that, we must decouple the two uses of
`fuchsia.ui.focus.FocusChainListenerRegistry` so that we can remove one
of the two, but keep the other.  This change does the decoupling.

Bug: 95901
Change-Id: Iaa73fa5c59feae4c02d908dfa3240cc855126485
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/666337
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>